updated app.js and PostPage.js
