+++ instagram/src/Components/SearchBar/SearchBar.css
@@ -27,6 +27,7 @@ h1 {
   font-size: 14px;
   text-align: center;
   padding: 2px;
+  font-family: FontAwesome;
 }
 
 .searchBar-icons {
