+++ package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "moment": "^2.23.0",
     "prop-types": "^15.6.2",
-    "react-simple-storage": "^1.4.1"
+    "react-simple-storage": "^1.4.1",
+    "styled-components": "^4.1.3"
   }
 }
